+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a problem where an error was thrown when a name from an importing module
   shadowed a nested name from the imported module (#802)
+- Fixed a problem where tests were ignored if they are not defined directly in
+  the main module - that is, they were imported (#1161)
 
 ### Security